+++ src/components/WorldManagement.tsx
@@ -1,18 +1,5 @@
-import React, { useState, useEffect } from "react";
-import {
-  Box,
-  Tabs,
-  Tab,
-  Typography,
-  Switch,
-  List,
-  ListItem,
-  ListItemText,
-  ListItemSecondaryAction,
-  Alert,
-  Snackbar,
-  CircularProgress,
-} from "@mui/material";
+import { useState, useEffect } from "react";
+import { Box, Tabs, Tab, Alert, Snackbar } from "@mui/material";
 import { useParams } from "react-router-dom";
 import axios from "axios";
 import { MinecraftWorld, Player } from "../types";
@@ -20,45 +7,7 @@ import { DatapackManagement } from "./DatapacKManagement";
 import { ServerPropertiesManagement } from "./ServerPropertiesManagement";
 import { ServerControlPanel } from "./ServerControlManagement";
 import { PlayerManagement } from "./PlayerManagement";
-
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-}
-
-const TabPanel = (props: TabPanelProps) => {
-  const { children, value, index, ...other } = props;
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`tabpanel-${index}`}
-      {...other}
-      style={{
-        height: "80vh",
-        position: "relative",
-        width: "100%",
-      }}
-    >
-      {value === index && (
-        <Box
-          sx={{
-            p: 3,
-            position: "absolute",
-            top: 0,
-            bottom: 0,
-            left: 0,
-            right: 0,
-            overflowX: "hidden",
-          }}
-        >
-          {children}
-        </Box>
-      )}
-    </div>
-  );
-};
+import { TabPanel } from "./TabPanel";
 
 export const WorldManagement = () => {
   const { worldId } = useParams<{ worldId: string }>();
